cli/requirements.txt
requests==2.31.0
qrcode==7.4.2
python-dotenv==1.0.0
Pillow==10.2.0
